requirements.txt
Django>=6.0
google-generativeai>=0.8.0




